Fix potential use after free, suggested by CodeQL
Commit message and changes are suggested by CodeQL.
 - https://github.com/nnstreamer/nnstreamer/security/code-scanning/2
 - https://github.com/nnstreamer/nnstreamer/security/code-scanning/3
 - https://github.com/nnstreamer/nnstreamer/security/code-scanning/4

To fix the potential use-after-free error, we need to ensure that the pointer trainer->prop.model_load_path is not accessed after it has been freed.
The best way to do this is to set the pointer to NULL immediately after freeing it.
This way, any subsequent access to the pointer can be checked against NULL, preventing the use of a dangling pointer.

Modify the function gst_tensor_trainer_set_model_load_path to set trainer->prop.model_load_path to NULL after calling g_free.

Signed-off-by: gichan2-jang <gichan2.jang@samsung.com>